requirements/base.txt
Django==3.2.21
django-environ==0.11.2
django-filter==23.4
django-model-utils==4.3.1
djangorestframework==3.14.0
mysqlclient==2.1.1